post EIP variance checker
this is a code designed to compare the variance of the pre and post-EIP gas price variance as a less-than-perfect way of approximating demand. The code works by combining the transaction and block datasets that barnabe gave us (txs_12965_12967 & bxs_12965_12967) to get all the required metrics and then rescaling the gas prices to a comparable metric to EIP using the method that I proposed in the chat earlier. the code then compares variance in two ways, normalizes the data (because of the discrepancy of the size of the two units) and compares 40000 randomly chosen observations from pre and post-transaction data and compares the variance, then performs a simulation where it grabs 500 random observations from each set and reports how often the variance is lower in the post-EIP gas prices. The code is very well documented, so any errors along the way should be easy to pick up. NOTE: this current version was performed with the non-ideal data, as a consequence of a lack of a running node at this point.  However, modifying it to run in such a way would be extremely simple once more data rolls in.